requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
python-multipart==0.0.6
pyway==0.3.14
python-jose==3.3.0
cryptography==41.0.1
passlib==1.7.4
bcrypt==4.0.1
motor == 3.1.2
pymongo == 4.4.0